move mappings to separate package
