+++ domain/content/repository/src/commonMain/kotlin/com/livefast/eattrash/raccoonforfriendica/domain/content/repository/DefaultTrendingRepository.kt
@@ -4,6 +4,8 @@ import com.livefast.eattrash.raccoonforfriendica.core.api.provider.ServiceProvid
 import com.livefast.eattrash.raccoonforfriendica.domain.content.data.LinkModel
 import com.livefast.eattrash.raccoonforfriendica.domain.content.data.TagModel
 import com.livefast.eattrash.raccoonforfriendica.domain.content.data.TimelineEntryModel
+import com.livefast.eattrash.raccoonforfriendica.domain.content.repository.utils.toModel
+import com.livefast.eattrash.raccoonforfriendica.domain.content.repository.utils.toModelWithReply
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.IO
 import kotlinx.coroutines.withContext
@@ -14,14 +16,14 @@ internal class DefaultTrendingRepository(
     override suspend fun getEntries(offset: Int): List<TimelineEntryModel> =
         withContext(Dispatchers.IO) {
             runCatching {
-            val response =
-                provider.trends
-                    .getStatuses(
-                        offset = offset,
-                        limit = DEFAULT_PAGE_SIZE,
-                    )
-            response.map { it.toModelWithReply() }
-        }.getOrElse { emptyList() }
+                val response =
+                    provider.trends
+                        .getStatuses(
+                            offset = offset,
+                            limit = DEFAULT_PAGE_SIZE,
+                        )
+                response.map { it.toModelWithReply() }
+            }.getOrElse { emptyList() }
         }
 
     override suspend fun getHashtags(offset: Int): List<TagModel> =
